Move LitPerturber into perturber.py
