+++ mart/attack/__init__.py
@@ -5,5 +5,6 @@ from .gain import *
 from .gradient_modifier import *
 from .initializer import *
 from .objective import *
+from .perturber import *
 from .projector import *
 from .threat_model import *
